requirements.txt
tensorflow==2.16.1
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
numpy==1.24.3
pandas==2.0.3
pillow==10.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pygame==2.5.2
gtts==2.4.0
scikit-learn==1.3.2